cleanup Updated synthetic action
