--- packages/client/src/v2-events/features/workqueues/EventOverview/components/EventHistory/EventHistoryDialog/components/DeclarationUpdate.tsx
@@ -1,57 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * OpenCRVS is also distributed under the terms of the Civil Registration
- * & Healthcare Disclaimer located at http://opencrvs.org/license.
- *
- * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
- */
-import React from 'react'
-import {
-  EventDocument,
-  EventStatus,
-  getStatusFromActions,
-  ValidatorContext
-} from '@opencrvs/commons/client'
-import { useEventConfiguration } from '@client/v2-events/features/events/useEventConfiguration'
-import { withSuspense } from '@client/v2-events/components/withSuspense'
-import { DeclarationComparisonTable } from '@client/v2-events/features/events/actions/correct/request/Summary/DeclarationComparisonTable'
-import { EventHistoryActionDocument } from '@client/v2-events/features/events/actions/correct/useActionForHistory'
-
-/**
- *
- * Displays diff between declaration and update of it.
- */
-function DeclarationUpdateComponent({
-  action,
-  fullEvent,
-  validatorContext
-}: {
-  action: EventHistoryActionDocument
-  fullEvent: EventDocument
-  validatorContext: ValidatorContext
-}) {
-  const { eventConfiguration } = useEventConfiguration(fullEvent.type)
-
-  const index = fullEvent.actions.findIndex((a) => a.id === action.id)
-  const previousStatus = getStatusFromActions(fullEvent.actions.slice(0, index))
-
-  // Comparison is not needed when there is nothing to compare to.
-  if (previousStatus === EventStatus.enum.CREATED) {
-    return null
-  }
-
-  return (
-    <DeclarationComparisonTable
-      action={action}
-      eventConfig={eventConfiguration}
-      fullEvent={fullEvent}
-      id={'declaration-update'}
-      validatorContext={validatorContext}
-    />
-  )
-}
-
-export const DeclarationUpdate = withSuspense(DeclarationUpdateComponent)
